+++ src/init.cc
@@ -28,6 +28,10 @@
 #include <unistd.h>
 #include "graph/topo.h"
 
+// [RCCL]
+#include "clique/CliqueManager.h"
+// [/RCCL]
+
 #define STR2(v) #v
 #define STR(v) STR2(v)
 
@@ -300,6 +304,7 @@ static ncclResult_t commFree(ncclComm_t comm) {
 }
 
 RCCL_PARAM(AllToAllDisable, "ALLTOALL_KERNEL_DISABLE", 1);
+RCCL_PARAM(ForceEnableClique, "FORCE_ENABLE_CLIQUE", 0);
 
 static ncclResult_t commAlloc(ncclComm_t* comret, int ndev, int rank) {
   if (ndev < 1) {
@@ -678,7 +683,10 @@ static ncclResult_t initTransportsRank(struct ncclComm* comm, ncclUniqueId* comm
   int nranks = comm->nRanks;
   uint64_t commHash = getHash(commId->internal, NCCL_UNIQUE_ID_BYTES);
   TRACE(NCCL_INIT, "comm %p, commHash %lx, rank %d nranks %d - BEGIN", comm, commHash, rank, nranks);
-  NCCLCHECK(bootstrapInit(commId, rank, nranks, &comm->bootstrap));
+  // [RCCL] Collect the PID of the root
+  int rootPid;
+  NCCLCHECK(bootstrapInit(commId, rank, nranks, &comm->bootstrap, &rootPid));
+  // [/RCCL]
 
   // AllGather1 - begin
   struct {
@@ -1016,7 +1024,52 @@ static ncclResult_t initTransportsRank(struct ncclComm* comm, ncclUniqueId* comm
       return ncclInternalError;
     }
     NCCLCHECK(ncclCommSetIntra(comm, intraRank, intraRanks, allGather1Data[intraRank0].comm));
+
+    { // [RCCL] Check if clique-based kernels can be enabled and initialize CliqueManager if so
+      CliqueManager::cliqueMode_t cliqueMode = CliqueManager::CLIQUE_DISABLED;
+      if (comm->localRanks == comm->nRanks)
+      {
+        // Check that all the GPUs have peer access to one another
+        bool hasPeerAccess = true;
+        for (int i = 0; i < nranks && hasPeerAccess; i++)
+        {
+          int cudaDev1 = allGather1Data[i].peerInfo.cudaDev;
+          for (int j = 0; j < nranks; j++)
+          {
+            if (i == j) continue;
+            int cudaDev2 = allGather1Data[j].peerInfo.cudaDev;
+            int p2p;
+            if (hipDeviceCanAccessPeer(&p2p, cudaDev1, cudaDev2) != hipSuccess || !p2p)
+            {
+              hasPeerAccess = false;
+              break;
+            }
+          }
+        }
+        if (hasPeerAccess)
+        {
+          if (intraRanks == nranks)
+            cliqueMode = CliqueManager::CLIQUE_SINGLE_PROCESS;
+          else
+            cliqueMode = CliqueManager::CLIQUE_SINGLE_NODE;
+        }
+
+        // For now, only enable clique-based kernels on CR8_G topologies, unless explicitly asked
+        if (!rcclParamForceEnableClique())
+        {
+          // Disable clique-kernel support if not on CR8 topology
+          if (!(comm->topo->nodes[NET].count == 0 && comm->topo->type == RCCL_TOPO_CR8G))
+          {
+            INFO(NCCL_INIT, "Disabling clique-based kernels due to topology (force enable with RCCL_FORCE_ENABLE_CLIQUE)");
+            cliqueMode = CliqueManager::CLIQUE_DISABLED;
+          }
+        }
+      }
+      comm->cliqueManager = new CliqueManager(rank, nranks, cliqueMode);
+      NCCLCHECK(comm->cliqueManager->Init(commId, rootPid));
+    } // [/RCCL]
   } while(0);
+
 
   // Done with AllGather1 data
   free(allGather1Data);
@@ -1146,6 +1199,10 @@ ncclResult_t ncclCommDestroy(ncclComm_t comm) {
     WARN("comm %p has already been destroyed", comm);
     return ncclInvalidArgument;
   }
+
+  // [RCCL] Delete CliqueManager if it exists
+  if (comm->cliqueManager) delete comm->cliqueManager;
+  // [/RCCL]
 
   return commDestroy(comm);
 }